+++ bitvector/bitvector.h
@@ -27,19 +27,19 @@ typedef struct b
     byte *bytes;
 }binary;
 
-size_t bit_count(bitvector *bv);
+uint bit_count(bitvector *bv);
 bitvector bv_create_empty();
-size_t bv_init(bitvector *bv);
-size_t bv_add(bitvector *bv, bool value);
-size_t bv_merge(bitvector *bv, bitvector *unionWith);
+int bv_init(bitvector *bv);
+uint bv_add(bitvector *bv, bool value);
+uint bv_merge(bitvector *bv, bitvector *unionWith);
 bool bv_remove_last(bitvector *bv);
 bool *bv_at(bitvector *bv, int index);
-size_t bv_clear(bitvector *bv);
+uint bv_clear(bitvector *bv);
 bool bv_delete(bitvector *bv);
 unsigned int bools2bits(bitvector *bv, binary *out);
 bool bits2bools(binary *b, bitvector *out);
-size_t write_binary_to_file(binary *b, char *file);
-size_t read_binary_from_file(char *file, binary *b);
+uint write_binary_to_file(binary *b, char *file);
+uint read_binary_from_file(char *file, binary *b);
 
 void printbitvector(bitvector *bv);
 void printbinary(binary *bin);
